+++ mlxtend/feature_selection/sequential_feature_selector.py
@@ -29,9 +29,14 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
     Parameters
     ----------
     estimator : scikit-learn classifier or regressor
-    k_features : int
+    k_features : int or tuple (new in 0.4.2) (default: 1)
         Number of features to select,
         where k_features < the full feature set.
+        New in 0.4.2: A tuple containing a min and max value can be provided,
+            and the SFS will consider return any feature combination between
+            min and max that scored highest in cross-validtion. For example,
+            the tuple (1, 4) will return any combination from
+            1 up to 4 features instead of a fixed number of features k.
     forward : bool (default: True)
         Forward selection if True,
         backward selection otherwise
@@ -93,7 +98,7 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
               'avg_score' (average cross-validation score)
 
     """
-    def __init__(self, estimator, k_features,
+    def __init__(self, estimator, k_features='best',
                  forward=True, floating=False,
                  print_progress=True, scoring='accuracy',
                  cv=5, skip_if_stuck=True, n_jobs=1,
@@ -135,19 +140,56 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
         self : object
 
         """
-        if X.shape[1] < self.k_features:
-            raise AttributeError('Features in X < k_features')
+
+        if not isinstance(self.k_features, int) and\
+                not isinstance(self.k_features, tuple):
+            raise AttributeError('k_features must be a positive integer'
+                                 ' or tuple')
+
+        if isinstance(self.k_features, int) and (self.k_features < 1 or
+                                                 self.k_features > X.shape[1]):
+            raise AttributeError('k_features must be a positive integer'
+                                 ' between 1 and X.shape[1], got %s'
+                                 % (self.k_features))
+
+        if isinstance(self.k_features, tuple):
+            if len(self.k_features) != 2:
+                raise AttributeError('k_features tuple must consist of 2'
+                                     ' elements a min and a max value.')
+
+            if self.k_features[0] not in range(1, X.shape[1] + 1):
+                raise AttributeError('k_features tuple min value must be in'
+                                     ' range(1, X.shape[1]+1).')
+
+            if self.k_features[1] not in range(1, X.shape[1] + 1):
+                raise AttributeError('k_features tuple max value must be in'
+                                     ' range(1, X.shape[1]+1).')
+
+            if self.k_features[0] > self.k_features[1]:
+                raise AttributeError('The min k_features value must be larger'
+                                     ' than the max k_features value.')
+
         if self.skip_if_stuck:
             sdq = deque(maxlen=4)
         else:
             sdq = deque(maxlen=0)
 
+        if isinstance(self.k_features, tuple):
+            select_in_range = True
+        else:
+            select_in_range = False
+            k_to_select = self.k_features
+
         self.subsets_ = {}
         orig_set = set(range(X.shape[1]))
         if self.forward:
+            if select_in_range:
+                k_to_select = self.k_features[1]
             k_idx = ()
             k = 0
         else:
+            if select_in_range:
+                k_to_select = self.k_features[0]
             k_idx = tuple(range(X.shape[1]))
             k = len(k_idx)
             k_score = self._calc_score(X, y, k_idx)
@@ -155,7 +197,7 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
                                 'cv_scores': k_score,
                                 'avg_score': k_score.mean()}
 
-        while k != self.k_features:
+        while k != k_to_select:
             prev_subset = set(k_idx)
             if self.forward:
                 k_idx, k_score, cv_scores = \
@@ -194,8 +236,17 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
 
             if self.print_progress:
                 sys.stderr.write('\rFeatures: %d/%d' % (
-                    len(k_idx), self.k_features))
+                    len(k_idx), k_to_select))
                 sys.stderr.flush()
+
+        if select_in_range:
+            max_score = -1
+            for k in self.subsets_:
+                if self.subsets_[k]['avg_score'] > max_score:
+                    max_score = self.subsets_[k]['avg_score']
+                    best_subset = k
+            k_score = max_score
+            k_idx = self.subsets_[best_subset]['feature_idx']
 
         self.k_feature_idx_ = k_idx
         self.k_score_ = k_score